[EI-254] [web] feat: Add SourceIndicatorCreateSliderGroup component for custom forecast indicator creation
