+++ apps/web/app/ui/components/domain/numerical-guidance/custom-forecast-indicator/custom-forecast-indicator-create-dialog-menu/source-indicator-create-slider-group.tsx
@@ -1,0 +1,31 @@
+import { useCreatingCustomForecastIndicator } from '@/app/business/hooks/numerical-guidance/custom-forecast-indicator/use-creating-custom-forecast-indicator.hook';
+import SourceIndicatorSlider, { SourceIndicatorInfo } from '../source-indicator-slider';
+
+export default function SourceIndicatorCreateSliderGroup() {
+  const { sourceIndicators, updateSourceIndicatorWeight, deleteSourceIndicator } = useCreatingCustomForecastIndicator();
+
+  const handleSourceIndicatorWeightChange = (item: SourceIndicatorInfo, value: number) => {
+    updateSourceIndicatorWeight(item.id, value);
+  };
+
+  const handleSourceIndicatorDelete = (item: SourceIndicatorInfo) => {
+    deleteSourceIndicator(item.id);
+  };
+
+  return (
+    <>
+      {sourceIndicators.map((sourceIndicator) => {
+        return (
+          <SourceIndicatorSlider
+            item={{
+              id: sourceIndicator.sourceIndicatorId,
+              ...sourceIndicator,
+            }}
+            onWeightChange={handleSourceIndicatorWeightChange}
+            onSourceIndicatorDelete={handleSourceIndicatorDelete}
+          />
+        );
+      })}
+    </>
+  );
+}
