+++ apps/web/app/ui/components/domain/numerical-guidance/custom-forecast-indicator/custom-forecast-indicator-create-dialog-menu/steps/select-source-indicator-step-dialog-menu.tsx
@@ -5,24 +5,10 @@ import { ListChildComponentProps } from 'react-window';
 import { Indicator } from '@/app/business/services/numerical-guidance/view-model/indicator-list/indicators/indicator.service';
 import SelectableItem from '@/app/ui/components/view/atom/selectable-item';
 import { useCreatingCustomForecastIndicator } from '@/app/business/hooks/numerical-guidance/custom-forecast-indicator/use-creating-custom-forecast-indicator.hook';
-import SourceIndicatorSlider, { SourceIndicatorInfo } from '../../source-indicator-slider';
+import SourceIndicatorCreateSliderGroup from '../source-indicator-create-slider-group';
 
 export default function SelectSourceIndicatorStepDialogMenu() {
-  const {
-    sourceIndicators,
-    addSourceIndicator,
-    deleteSourceIndicator,
-    includeSourceIndicator,
-    updateSourceIndicatorWeight,
-  } = useCreatingCustomForecastIndicator();
-
-  const handleSourceIndicatorWeightChange = (item: SourceIndicatorInfo, value: number) => {
-    updateSourceIndicatorWeight(item.id, value);
-  };
-
-  const handleSourceIndicatorDelete = (item: SourceIndicatorInfo) => {
-    deleteSourceIndicator(item.id);
-  };
+  const { addSourceIndicator, deleteSourceIndicator, includeSourceIndicator } = useCreatingCustomForecastIndicator();
 
   const render = ({ index, style, data }: ListChildComponentProps<Indicator[]>) => {
     const indicator = data[index];
@@ -60,18 +46,7 @@ export default function SelectSourceIndicatorStepDialogMenu() {
       </DialogMenu.Content>
       <DialogMenu.Content>
         <div className="py-1 text-xs font-bold">가중치</div>
-        {sourceIndicators.map((sourceIndicator) => {
-          return (
-            <SourceIndicatorSlider
-              item={{
-                id: sourceIndicator.sourceIndicatorId,
-                ...sourceIndicator,
-              }}
-              onWeightChange={handleSourceIndicatorWeightChange}
-              onSourceIndicatorDelete={handleSourceIndicatorDelete}
-            />
-          );
-        })}
+        <SourceIndicatorCreateSliderGroup />
       </DialogMenu.Content>
     </>
   );
